requirements.txt
notion-client==2.5.0
requests
ruff
python-dotenv